pyproject.toml
[project]
name = "app_template"
version = "0.1.0"
description = "Python app scaffolded from templates"
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
    "python-dotenv",
    "pytest",
    "gradio",
    "openai>=1.0.0",
    "boto3"
]

[tool.uv]
# left empty unless you add uv-specific config
